fix: Complete i32 type conversion for SQLite INT4 compatibility
Changed all integer ID fields and route parameters to i32 to match
SQLite's INT4 storage type. This resolves the type mismatch error:
"Rust type i64 is not compatible with SQL type INT4"

Changes:
- routes.rs: All route parameters now use i32
- models.rs: All struct ID fields now use i32
- guards.rs: Auth guard uses i32 for store IDs
- ai.rs: AI response structs use i32 for IDs

+++ server/src/t2/ai.rs
@@ -267,7 +267,7 @@ match_score от 0 до 100 - насколько товар соответств
 
     #[derive(Deserialize)]
     struct AIRecommendation {
-        product_id: i64,
+        product_id: i32,
         price_category: String,
         match_score: i32,
         match_reasons: Vec<String>,
@@ -366,7 +366,7 @@ pub async fn recommend_accessories(
 
     #[derive(Deserialize)]
     struct AIAccessory {
-        product_id: i64,
+        product_id: i32,
         reason: String,
         benefit: String,
     }
@@ -459,7 +459,7 @@ pub async fn recommend_tariffs(
 
     #[derive(Deserialize)]
     struct AITariff {
-        tariff_id: i64,
+        tariff_id: i32,
         recommendation: String,
     }
