+++ server/src/t2/models.rs
@@ -3,8 +3,7 @@ use sqlx::FromRow;
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Store {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
+    pub id: i32,
     pub name: String,
     pub address: String,
     pub admin_code: String,
@@ -14,10 +13,8 @@ pub struct T2Store {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Employee {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub store_id: i64,
+    pub id: i32,
+    pub store_id: i32,
     pub name: String,
     pub code: String,
     pub is_admin: bool,
@@ -26,8 +23,8 @@ pub struct T2Employee {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct T2EmployeeWithStores {
-    pub id: i64,
-    pub store_id: i64,
+    pub id: i32,
+    pub store_id: i32,
     pub name: String,
     pub code: String,
     pub is_admin: bool,
@@ -37,8 +34,7 @@ pub struct T2EmployeeWithStores {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Category {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
+    pub id: i32,
     pub name: String,
     pub icon: Option<String>,
     pub created_at: String,
@@ -46,32 +42,25 @@ pub struct T2Category {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Tag {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub store_id: i64,
+    pub id: i32,
+    pub store_id: i32,
     pub name: String,
     pub color: String,
     pub description: Option<String>,
-    #[sqlx(try_from = "i32")]
-    pub priority: i64,
+    pub priority: i32,
     pub created_at: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Product {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub store_id: i64,
-    #[sqlx(try_from = "i32")]
-    pub category_id: i64,
+    pub id: i32,
+    pub store_id: i32,
+    pub category_id: i32,
     pub name: String,
     pub brand: Option<String>,
     pub model: Option<String>,
     pub price: f64,
-    #[sqlx(try_from = "i32")]
-    pub quantity: i64,
+    pub quantity: i32,
     pub image_url: Option<String>,
     pub created_at: String,
     pub updated_at: String,
@@ -79,15 +68,15 @@ pub struct T2Product {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct T2ProductWithDetails {
-    pub id: i64,
-    pub store_id: i64,
-    pub category_id: i64,
+    pub id: i32,
+    pub store_id: i32,
+    pub category_id: i32,
     pub category_name: String,
     pub name: String,
     pub brand: Option<String>,
     pub model: Option<String>,
     pub price: f64,
-    pub quantity: i64,
+    pub quantity: i32,
     pub image_url: Option<String>,
     pub specs: Vec<T2ProductSpec>,
     pub tags: Vec<T2Tag>,
@@ -97,20 +86,16 @@ pub struct T2ProductWithDetails {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2ProductSpec {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub product_id: i64,
+    pub id: i32,
+    pub product_id: i32,
     pub spec_name: String,
     pub spec_value: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Tariff {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub store_id: i64,
+    pub id: i32,
+    pub store_id: i32,
     pub name: String,
     pub price: f64,
     pub minutes: Option<i32>,
@@ -128,10 +113,8 @@ pub struct T2Tariff {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Service {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub store_id: i64,
+    pub id: i32,
+    pub store_id: i32,
     pub name: String,
     pub price: f64,
     pub description: Option<String>,
@@ -141,12 +124,9 @@ pub struct T2Service {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Sale {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub store_id: i64,
-    #[sqlx(try_from = "i32")]
-    pub employee_id: i64,
+    pub id: i32,
+    pub store_id: i32,
+    pub employee_id: i32,
     pub customer_request: Option<String>,
     pub customer_audio_url: Option<String>,
     pub total_amount: f64,
@@ -156,10 +136,10 @@ pub struct T2Sale {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct T2SaleWithDetails {
-    pub id: i64,
-    pub store_id: i64,
+    pub id: i32,
+    pub store_id: i32,
     pub store_name: String,
-    pub employee_id: i64,
+    pub employee_id: i32,
     pub employee_name: String,
     pub customer_request: Option<String>,
     pub customer_audio_url: Option<String>,
@@ -171,27 +151,21 @@ pub struct T2SaleWithDetails {
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2SaleItem {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub sale_id: i64,
+    pub id: i32,
+    pub sale_id: i32,
     pub item_type: String,
-    #[sqlx(try_from = "i32")]
-    pub item_id: i64,
+    pub item_id: i32,
     pub item_name: String,
     pub item_details: Option<String>,
     pub price: f64,
-    #[sqlx(try_from = "i32")]
-    pub quantity: i64,
+    pub quantity: i32,
     pub created_at: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct T2Session {
-    #[sqlx(try_from = "i32")]
-    pub id: i64,
-    #[sqlx(try_from = "i32")]
-    pub employee_id: i64,
+    pub id: i32,
+    pub employee_id: i32,
     pub token: String,
     pub expires_at: String,
     pub created_at: String,
@@ -218,13 +192,13 @@ pub struct CreateStoreRequest {
 
 #[derive(Debug, Deserialize)]
 pub struct CreateEmployeeRequest {
-    pub store_id: i64,
+    pub store_id: i32,
     pub name: String,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct CreateProductRequest {
-    pub category_id: i64,
+    pub category_id: i32,
     pub name: String,
     pub brand: Option<String>,
     pub model: Option<String>,
@@ -232,7 +206,7 @@ pub struct CreateProductRequest {
     pub quantity: Option<i32>,
     pub image_url: Option<String>,
     pub specs: Vec<ProductSpecInput>,
-    pub tag_ids: Vec<i64>,
+    pub tag_ids: Vec<i32>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -250,7 +224,7 @@ pub struct UpdateProductRequest {
     pub quantity: Option<i32>,
     pub image_url: Option<String>,
     pub specs: Option<Vec<ProductSpecInput>>,
-    pub tag_ids: Option<Vec<i64>>,
+    pub tag_ids: Option<Vec<i32>>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -376,12 +350,12 @@ pub struct CreateSaleRequest {
 #[derive(Debug, Deserialize)]
 pub struct SaleItemInput {
     pub item_type: String,
-    pub item_id: i64,
+    pub item_id: i32,
     pub quantity: Option<i32>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct SearchProductsRequest {
     pub query: String,
-    pub category_id: Option<i64>,
+    pub category_id: Option<i32>,
 }
